Phase AE: continuous fuzzing of every hypervisor device-emulation path
Adds deterministic, fixed-seed total-no-panic fuzz coverage over every
guest-untrusted surface, wired into CI as standing gates, plus a growing
committed corpus. This is the fuzzing half of the security-hardening push
(Phase AD Miri + ASan already landed).

New fuzz targets (all assert total no-panic on hostile input):
- vmx.rs guest_controlled_exit_decodes_are_total_and_fail_closed: 1M hostile
  VMX guest-control decode + EPT-violation qualification inputs (reserved size
  encodings + every reason path exercised, fail-closed).
- vdev.rs guest_device_io_never_panics_on_hostile_port_sequences: 500k hostile
  (port,value) sweeps over the whole guest port surface DeviceSet
  (PIC/UART/PIT/RTC/PCI/virtio/UHCI/SB16), interleaving reads/writes on live
  device state so mid-ICW / DLAB states are reached.
- virtio.rs hostile_descriptor_tables_and_sectors_never_panic: 100k hostile
  descriptor tables + avail-ring idx growth against VirtioBlk + MemStore.
- hardening_fuzz.rs tls_record_framing_and_server_hello_never_panic: 200k
  hostile TLS record-framing + ServerHello parses (both accepted paths hit).
- fuzz_corpus.rs corpus_driven_fuzz_is_total_and_grows: corpus-driven harness,
  11 targets (ethernet/arp/ipv4/udp/tcp/tls_record/tls_server_hello/vmx_io/
  vmx_ept/vmx_opcode/vmx_reason), FNV-1a content-hashed seeds in
  aegis-kernel/fuzz-corpus/, growth gated on AEGIS_FUZZ_GROW=1, budget
  AEGIS_FUZZ_ITERS, single sequential test (corpus writes not thread-safe),
  #cfg_attr(miri,ignore).

CI (.github/workflows/test.yml): fuzz-push (bounded read-only release pass on
every push) + fuzz-nightly (long growth pass, commits corpus growth back).

Sweep budgets scale 1/4 in debug builds via sweep_iters() so the every-push
gate stays a bounded few minutes; full budget runs in release / nightly.
Made Rng/SEED/no_panic and the VMX decoders pub(crate) so the new suites can
reuse them.

Raw verification (clean lockfiles, all release except as noted):
  kernel plain:      test result: ok. 792 passed; 0 failed
  kernel vmx-demo:   test result: ok. 795 passed; 0 failed
  workspace:         136 passed; 0 failed
  bootloader:        22 passed; 0 failed
  verify-security-audit.sh: OK (live totals 950 match Docs row)
  reachable-authority audit: 0 violations, 0 warnings
  committed corpus:  2049 seeds, harness 0 panics
  debug suite:       792 passed; 0 failed (sweep_iters keeps it ~6 min)

+++ aegis-kernel/src/syscall.rs
@@ -103,7 +103,7 @@ fn dispatch_impl(num: u64, arg1: u64, arg2: u64, arg3: u64, arg4: u64) -> i64 {
             // Write: print arg2 bytes from the buffer at arg1 to COM1
             // (and mirror them to the VGA text console). The whole range must
             // pass the user-pointer gate first.
-            let len = clamp_write_len(arg2) as usize;
+            let len = clamp_write_len(arg2);
             if !crate::user_ptr::validate_range(
                 crate::user_ptr::current_user_pml4(),
                 arg1,
